% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSgroups.R
\name{addObject3js}
\alias{addObject3js}
\title{Generate a new object ID}
\usage{
addObject3js(data3js, object, number_of_ids = 1)
}
\arguments{
\item{data3js}{The r3js data object}

\item{num}{The number of new IDs to generate}
}
\value{
Returns the update r3js data object
}
\description{
Generate a new object ID for an r3js plot.
}
