% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSobjects.R
\name{arrow3js}
\alias{arrow3js}
\title{Add an arrow to a data3js object}
\usage{
arrow3js(
  data3js,
  x,
  y,
  z,
  lwd = 1,
  arrowhead_width = 0.2,
  arrowhead_length = 0.5,
  col,
  mat = "basic",
  ...
)
}
\arguments{
\item{data3js}{The data3js object}

\item{x}{x coordinates (from then to)}

\item{y}{y coordinates (from then to)}

\item{z}{z coordinates (from then to)}

\item{lwd}{line width}

\item{arrowhead_width}{arrowhead width}

\item{arrowhead_length}{arrowhead length}

\item{col}{color}

\item{mat}{material (see `material3js()`)}

\item{...}{other arguments to pass to `material3js()`}
}
\description{
This is not yet implemented
}
