% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSsetup.R
\name{axis3js}
\alias{axis3js}
\title{Add axes to an r3js plot}
\usage{
axis3js(data3js, side, loc = "front", lims = "--", at = NULL,
  labels = NULL, lwd = 1, mat = "basic", cex.axis = 1, grid = TRUE,
  grid_style = list(lwd = 1, mat = "basic", col = "grey80"), ...)
}
\arguments{
\item{at}{}
}
\description{
Add axes to an r3js plot
}
