% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSplotting.R
\name{axislines3js}
\alias{axislines3js}
\title{Add a grid to an r3js plot}
\usage{
axislines3js(data3js, x, y, z, ax, lwd = 0.9, col = "grey80", ...)
}
\arguments{
\item{z}{}
}
\description{
Add a grid parallel to the axis to an r3js plot. This is mostly intended for
adding grid lines to an axis.
}
