% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSsetup.R
\name{background3js}
\alias{background3js}
\title{Set the plot background color}
\usage{
background3js(data3js, col)
}
\arguments{
\item{data3js}{The data3js object}

\item{col}{The background color}
}
\value{
Returns an updated data3js object
}
\description{
Set the plot background color
}
