% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSsetup.R
\name{box3js}
\alias{box3js}
\title{Add a box to an r3js plot}
\usage{
box3js(
  data3js,
  sides = c("x", "y", "z"),
  dynamic = TRUE,
  col = "grey80",
  geometry = FALSE,
  renderOrder = 1,
  ...
)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
Add a box to an r3js plot
}
