% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSproperties.R
\name{col_opacity}
\alias{col_opacity}
\title{Internal function to get opacity from a color specification}
\usage{
col_opacity(col)
}
\description{
Internal function to get opacity from a color specification
}
