% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSproperties.R
\name{convertCol3js}
\alias{convertCol3js}
\title{Internal function to convert color to an rgb list}
\usage{
convertCol3js(col)
}
\arguments{
\item{col}{The color to convert in a format R understands}
}
\value{
Returns a list with red, green and blue values
}
\description{
Internal function to convert color to an rgb list
}
