% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSsetup.R
\name{export3js}
\alias{export3js}
\title{Export an r3js plot}
\usage{
export3js(data3js, file, title = "r3js plot", ...)
}
\arguments{
\item{data3js}{The}

\item{...}{}
}
\description{
Exports a r3js plot to an external html file.
}
