% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSsetup.R
\name{export3jsWidget}
\alias{export3jsWidget}
\title{Export an r3js html widget}
\usage{
export3jsWidget(widget, file, title = "r3js plot")
}
\arguments{
\item{widget}{The}

\item{...}{}
}
\description{
Export an r3js html widget to an external html file.
}
