% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSlines.R
\name{gllines3js}
\alias{gllines3js}
\title{Add a non-geometric line to a data3js object}
\usage{
gllines3js(data3js, x, y, z, lwd = 1, col = "black", segments = FALSE, ...)
}
\arguments{
\item{data3js}{The data3js object}

\item{x}{x coords}

\item{y}{y coords}

\item{z}{z coords}

\item{lwd}{line width}

\item{col}{line color}

\item{segments}{Should the line be interpreted as segments, e.g. a dashed line}

\item{...}{Additional arguments to pass to \code{material3js()}}
}
\description{
This underlies the \code{lines3js()} function
}
