% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSbaseobjects.R
\name{glpoints3js}
\alias{glpoints3js}
\title{Add webgl points to a plot}
\usage{
glpoints3js(data3js, x, y, z, size, col, pch, highlight, ...)
}
\arguments{
\item{data3js}{The r3js data object}

\item{x}{x coordinates}

\item{y}{y coordinates}

\item{z}{z coordinates}

\item{size}{point sizes}

\item{col}{point colors}

\item{pch}{plotting characters}

\item{highlight}{highlight}

\item{...}{further parameters to pass to material3js}
}
\value{
Returns and updated r3js plotting object
}
\description{
Adds points to a plot represented as webgl points rather than geometries. They are a much more
basic representation, will not obey lighting rules and overlay rather than intersect but are
orders of magnitude more highly performent.
}
