% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSgrid.R
\name{grid3js}
\alias{grid3js}
\title{Add axis grids to an data3js object}
\usage{
grid3js(
  data3js,
  sides = c("x", "y", "z"),
  axes = c("x", "y", "z"),
  at = NULL,
  dynamic = TRUE,
  col = "grey95",
  geometry = FALSE,
  ...
)
}
\arguments{
\item{data3js}{The data3js object}

\item{sides}{The axis sides to show the box, any combination of "x", "y" or "z"}

\item{axes}{Axes for which to draw the grid lines}

\item{at}{Where to draw grid lines along the axis}

\item{dynamic}{Should edges of the box closest to the viewer hide themselves automatically}

\item{col}{Grid line color}

\item{geometry}{Should the lines be rendered as a physical geometry in the scene (see `lines3js()`)}

\item{...}{Other arguments to pass to `material3js()`}
}
\description{
This is used for example by `plot3js()` to add axis grids to a plot
these show along the faces of the plotting box, indicating axis ticks.
}
