% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSplotting.R
\name{highlight3js}
\alias{highlight3js}
\title{Add an r3js highlight object}
\usage{
highlight3js(data3js, highlight)
}
\arguments{
\item{highlight}{}
}
\value{

}
\description{
Add an r3js highlight object
}
