% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSgroups.R
\name{lastID}
\alias{lastID}
\title{Find the ID of the last object(s) added}
\usage{
lastID(data3js)
}
\arguments{
\item{data3js}{The r3js data object}
}
\value{
Returns a vector of IDs for the last objects added
}
\description{
Find the ID of the last object(s) added to an r3js plot.
}
