% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSbaseobjects.R
\name{light3js}
\alias{light3js}
\title{Add a light source to a plot}
\usage{
light3js(data3js, position = NULL)
}
\arguments{
\item{position}{Position of the light source}
}
\value{
Returns the updated r3js object
}
\description{
Add a light source to a plot
}
