% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSplotting.R
\name{line3js}
\alias{line3js}
\title{Add a line to a 3js object}
\usage{
line3js(data3js, x0, y0, z0, x1, y1, z1, lwd = 1, col = "black",
  mat = "basic", ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Add a line to a 3js object
}
