% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSsetup.R
\name{mtext3js}
\alias{mtext3js}
\title{Add text to a margin}
\usage{
mtext3js(data3js, text, side, line = 0, at = 0.5, cornerside = "f",
  ...)
}
\arguments{
\item{...}{}
}
\description{
Add text to a margin
}
