% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSsetup.R
\name{plot3js.new}
\alias{plot3js.new}
\title{Setup a new r3js plot}
\usage{
plot3js.new(background = "#ffffff")
}
\arguments{
\item{background}{Background color to use}
}
\value{
Returns a new data3js plotting object
}
\description{
This function sets up a new r3js plot and returns an r3js plotting object
that can later be added to using other functions such as \code{points3js()}
and \code{lines3js()} etc.  It is in many ways equivalent to the \code{plot.new()}
command.
}
