% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSplotting.R
\name{plot3js}
\alias{plot3js}
\title{Generic r3js plotting function}
\usage{
plot3js(x, y, z, aspect = c(1, 1, 1), fixed_aspect = FALSE,
  show_plot = TRUE, label_axes = TRUE, show_grid = TRUE, ...)
}
\arguments{
\item{...}{}
}
\description{
Generic r3js plotting function
}
