% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSsetup.R
\name{plot3js.window}
\alias{plot3js.window}
\title{Set axis limits for a 3js plot}
\usage{
plot3js.window(data3js, xlim, ylim, zlim, aspect = c(1, 1, 1))
}
\arguments{
\item{zlim}{}
}
\value{

}
\description{
Set axis limits for a 3js plot
}
