% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSaxes.R
\name{pretty_axis}
\alias{pretty_axis}
\title{Function to generate pretty axis labels that don't go outside range of x}
\usage{
pretty_axis(x, n = 5, include_lims = TRUE)
}
\arguments{
\item{x}{A numeric vector}

\item{n}{Desired number of break points}

\item{include_lims}{Should axis limits be included or excluded}
}
\description{
Function to generate pretty axis labels that don't go outside range of x
}
