% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSmethods.R
\name{print.data3js}
\alias{print.data3js}
\title{Printing r3js data information}
\usage{
\method{print}{data3js}(data3js)
}
\arguments{
\item{data3js}{The r3js data object}
}
\description{
Prints information about an r3js data object.
}
