% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r3js.R
\name{r3js}
\alias{r3js}
\title{Plot r3js data}
\usage{
r3js(data3js, width = NULL, height = NULL, elementId = NULL,
  rotation = c(-45, 0, 0), zoom = 1, translation = c(0, 0, 0),
  show_rotation = FALSE)
}
\description{
Creates an html widget from an r3js data object.
}
