% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSsetup.R
\name{save3jsWidget}
\alias{save3jsWidget}
\title{Save an r3js widget to an HTML file}
\usage{
save3jsWidget(widget, file, title = "r3js plot", selfcontained = TRUE,
  libdir = NULL, ...)
}
\arguments{
\item{widget}{Widget to save}

\item{file}{File to save HTML into}

\item{title}{Text to use as the title of the generated page}

\item{selfcontained}{Whether to save the HTML as a single self-contained file (with external resources base64 encoded) or a file with external resources placed in an adjacent directory}

\item{libdir}{Directory to copy HTML dependencies into (defaults to filename_files)}

\item{...}{Further arguments to pass to \code{\link[htmlwidgets]{saveWidget}}}
}
\description{
Save a rendered r3js widget to an HTML file (e.g. for sharing with others). This is mostly a wrapper for \code{\link[htmlwidgets]{saveWidget}}.
}
