% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSlines.R
\name{segments3js}
\alias{segments3js}
\title{Add lines segments a 3js object}
\usage{
segments3js(
  data3js,
  x,
  y,
  z,
  lwd = 1,
  col = "black",
  highlight,
  geometry = FALSE,
  ...
)
}
\arguments{
\item{data3js}{The data3js object}

\item{x}{x coords}

\item{y}{y coords}

\item{z}{z coords}

\item{lwd}{line width}

\item{col}{line color}

\item{highlight}{highlight characteristics (see `highlight3ks()`)}

\item{geometry}{logical, should the lines be rendered as a physical geometries}

\item{...}{further parameters to pass to `material3js()`}
}
\description{
Add lines segments a 3js object
}
