% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSbaseobjects.R, R/3JSplotting.R
\name{shape3js}
\alias{shape3js}
\title{Add a generic shape to an 3js plot}
\usage{
shape3js(data3js, vertices, faces, col = "black", highlight,
  label = NULL, toggle = NULL, ...)

shape3js(data3js, vertices, faces, col = "black", highlight,
  label = NULL, toggle = NULL, ...)
}
\arguments{
\item{...}{}

\item{...}{}
}
\description{
Add a generic shape to an 3js plot

Add a generic shape to an 3js plot
}
