% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSgrid.R
\name{sidegrid3js}
\alias{sidegrid3js}
\title{Add a side grid to the plot}
\usage{
sidegrid3js(
  data3js,
  axis,
  side,
  at = NULL,
  col = "grey80",
  dynamic = FALSE,
  geometry = FALSE,
  ...
)
}
\arguments{
\item{data3js}{The data3js object}

\item{axis}{Axis for which to draw lines, one of "x", "y" or "z"}

\item{side}{The side of the grid, one of "x", "y" or "z"}

\item{at}{Where to draw grid lines, defaults to the current axis tick marks}

\item{col}{Color of the grid lines}

\item{dynamic}{Should edges of the box closest to the viewer hide themselves automatically}

\item{geometry}{Should the lines be rendered as a physical geometry in the scene (see \code{lines3js()})}

\item{...}{Other arguments to pass to \code{material3js()}}
}
\description{
This is used by the function grid3js to add a grid on the faces of the plot region,
marking for example axis points.
}
