% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSutils.R
\name{snapshot3js}
\alias{snapshot3js}
\title{Create a snapshot of an r3js plot}
\usage{
snapshot3js(data3js, filename, ...)
}
\arguments{
\item{data3js}{The r3js data object}

\item{filename}{File to save the snapshot to}

\item{...}{Further parameters to pass to r3js}
}
\description{
Create a snapshot of an r3js plot
}
