% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSobjects.R
\name{sphere3js}
\alias{sphere3js}
\title{Add a sphere of defined radius to a data3js object}
\usage{
sphere3js(data3js, x, y, z, radius, col, highlight, ...)
}
\arguments{
\item{data3js}{The data3js object}

\item{x}{x coordinate of the sphere center}

\item{y}{y coordinate of the sphere center}

\item{z}{z coordinate of the sphere center}

\item{radius}{sphere radius}

\item{col}{color}

\item{highlight}{highlight attributes (see `highlight3js()`)}

\item{...}{other arguments to pass to `material3js()`}
}
\description{
Unlike points3js, where geometric points can also be represented as
spheres, this adds sphere that is sized with respect to the actual
dimensions of the plotting space (and so if aspect ratios differ for
each axis may not actually appear sphere-like!).
}
