% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSbaseobjects.R
\name{string3js}
\alias{string3js}
\title{Add a single text string to a 3js plot}
\usage{
string3js(
  data3js,
  x,
  y,
  z,
  text,
  size = 1,
  col = "inherit",
  type = "geometry",
  label = NULL,
  toggle = NULL,
  alignment = "center",
  offset = c(0, 0),
  style = list(),
  ...
)
}
\arguments{
\item{pos}{}
}
\description{
Add a single text string to a 3js plot
}
