% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{style3js}
\alias{style3js}
\title{Convert arguments to a style setting}
\usage{
style3js(...)
}
\arguments{
\item{...}{The list of styles and values}
}
\value{
Returns a list of styles marked with `jsonlite::unbox()`
}
\description{
An internal function for converting a string of arguments into a set of css styles for an object.
}
