% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSplotting.R
\name{surface3js}
\alias{surface3js}
\title{Add a surface to a 3js object}
\usage{
surface3js(data3js, x, y, z, col, mat, wireframe = FALSE, highlight, ...)
}
\arguments{
\item{...}{}
}
\description{
Add a surface to a 3js object
}
