% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teapot}
\alias{teapot}
\title{Utah Teapot}
\format{
A list with components vertices and edges. vertices is a 1976 by 3
numeric matrix of the coordinates of the vertices. edges is a 3751 by 3
integer matrix of the indices of the triangles.
}
\source{
Taken from the \code{misc3d} package, in turn from the netCDF file that
was at one time made available by Dave Forrest at
\url{http://www.maplepark.com/~drf5n/extras/teapot.nc}.
}
\usage{
teapot
}
\description{
The Utah teapot is a classic computer graphics example. This data set
contains a representation in terms of triangles. This is taken from the
\code{misc3d} package.
}
\keyword{datasets}
