% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSplotting.R
\name{text3js}
\alias{text3js}
\title{Add text to a 3js plot}
\usage{
text3js(data3js, x, y, z, text, size = 1, col = "black", label = NULL,
  toggle = NULL, type = "geometry", pos = c(0.5, 0.5), offset = c(0, 0),
  ...)
}
\arguments{
\item{...}{}
}
\description{
Add text to a 3js plot
}
