% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSobjects.R
\name{triangle3js}
\alias{triangle3js}
\title{Add a triangle to a data3js object}
\usage{
triangle3js(data3js, vertices, col = "black", highlight, ...)
}
\arguments{
\item{data3js}{The data3js object}

\item{vertices}{An nx3 matrix of triangle vertices}

\item{col}{Color of the triangle}

\item{highlight}{highlight attributes (see `highlight3js()`)}

\item{...}{Additional attributes to pass to `material3js()`}
}
\description{
Add a triangle to a data3js object
}
