% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSgroups.R
\name{addObject3js}
\alias{addObject3js}
\title{Add a plot object to a data3js object}
\usage{
addObject3js(data3js, object, number_of_ids = 1)
}
\arguments{
\item{data3js}{The r3js data object}

\item{object}{The plot object data to add}

\item{number_of_ids}{The number of ids to assign to the object, e.g. 10 when adding 10 points.}
}
\value{
Returns the updated data3js object
}
\description{
This is used internally to add an object to a data3js object, it assigns the
object an id and adds a record to the data3js object.
}
