% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSbaseobjects.R
\name{arrow3js}
\alias{arrow3js}
\title{Add an arrow to a 3js object}
\usage{
arrow3js(
  data3js,
  x,
  y,
  z,
  lwd = 1,
  arrowhead_width = 0.2,
  arrowhead_length = 0.5,
  col,
  mat = "basic",
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Add an arrow to a 3js object
}
