% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSsetup.R
\name{axis3js}
\alias{axis3js}
\title{Add axes to an r3js plot}
\usage{
axis3js(data3js, side, at = NULL, labels = NULL, cornerside = "f",
  ...)
}
\arguments{
\item{at}{}
}
\description{
Add axes to an r3js plot
}
