% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSsetup.R
\name{background3js}
\alias{background3js}
\title{Set the background color of an r3js plot}
\usage{
background3js(data3js, col)
}
\arguments{
\item{data3js}{The plotting object}

\item{col}{The background color}
}
\description{
Set the background color of an r3js plot
}
