% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSsetup.R
\name{box3js}
\alias{box3js}
\title{Add a box to an r3js plot}
\usage{
box3js(
  data3js,
  sides = c("x", "y", "z"),
  dynamic = TRUE,
  col = "grey80",
  geometry = FALSE,
  renderOrder = 1,
  ...
)
}
\arguments{
\item{data3js}{The data3js object}

\item{sides}{The axis side to show the box, any combination of "x", "y" or "z"}

\item{dynamic}{Should edges of the box closest to the viewer hide themselves automatically}

\item{col}{Box color}

\item{geometry}{Should the box be rendered as a physical geometry in the scene (see \code{lines3js()})}

\item{renderOrder}{The render order for the box, defaults to 1}

\item{...}{Other arguments to pass to \code{material3js()}}
}
\description{
Add a box to an r3js plot
}
\examples{
p <- plot3js.new()
box3js(p)

}
