% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSproperties.R
\name{clippingPlane3js}
\alias{clippingPlane3js}
\title{Create a clipping plane object}
\usage{
clippingPlane3js(coplanarPoints = NULL)
}
\arguments{
\item{coplanarPoints}{A matrix of 3 points coplanar to the plane, each row is a point, cols are coordinates}
}
\value{
Returns and r3js clipping plane object
}
\description{
This function can be used to create a clipping plane that can then be applied
to individual objects in a plot
}
