% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSpoints.R
\name{geopoint3js}
\alias{geopoint3js}
\title{Add a geometric point to a data3js object}
\usage{
geopoint3js(
  data3js,
  x,
  y,
  z,
  size,
  col = "black",
  fill = col,
  shape = "sphere",
  highlight,
  ...
)
}
\arguments{
\item{data3js}{The data3js object}

\item{x}{x coordinate}

\item{y}{y coordinate}

\item{z}{z coordinate}

\item{size}{point size}

\item{col}{point color}

\item{fill}{point fill color}

\item{shape}{point shape}

\item{highlight}{point highlight appearance (see \code{highlight3js()})}

\item{...}{other attributes to pass to \code{material3js()}}
}
\description{
This function underlies the points3js function when the geometry argument is set to TRUE.
The point generated is included a 'physical' geometry, for example a sphere if shape = "circle",
it will overlap with other points and shade according to the light source more realistically
but is \emph{much} less computationally efficient than glpoints, which uses a single point plus
different shader instructions depending on the point type.
}
