% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSbaseobjects.R
\name{gllines3js}
\alias{gllines3js}
\title{Insert a line into an r3js plot}
\usage{
gllines3js(data3js, x, y, z, lwd = 1, col = "black",
  segments = FALSE, ...)
}
\arguments{
\item{...}{}
}
\description{
Insert a line into an r3js plot
}
