% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSpoints.R
\name{glpoints3js}
\alias{glpoints3js}
\title{Add webgl points to a data3js object}
\usage{
glpoints3js(
  data3js,
  x,
  y,
  z,
  size = 1,
  col = "black",
  pch = 16,
  highlight,
  ...
)
}
\arguments{
\item{data3js}{The data3js object}

\item{x}{x coordinates}

\item{y}{y coordinates}

\item{z}{z coordinates}

\item{size}{point sizes}

\item{col}{point colors}

\item{pch}{plotting characters}

\item{highlight}{highlight characteristics (see `highlight3ks()`)}

\item{...}{further parameters to pass to `material3js()`}
}
\description{
Adds points to a plot represented as webgl points rather than geometries. They are a much more
basic representation, will not obey lighting rules and overlay rather than intersect but are
orders of more efficient to render.
}
