% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSsetup.R
\name{grid3js}
\alias{grid3js}
\title{Add axis grids to an r3js plot}
\usage{
grid3js(
  data3js,
  sides = c("x", "y", "z"),
  axes = c("x", "y", "z"),
  at = NULL,
  dynamic = TRUE,
  col = "grey95",
  geometry = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Add axis grids to an r3js plot
}
