% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSgroups.R
\name{group3js}
\alias{group3js}
\title{Start a new r3js object group}
\usage{
group3js(data3js, objectIDs, groupIDs = objectIDs)
}
\arguments{
\item{data3js}{The r3js data object}

\item{objectIDs}{IDs for each object you want to apply the group to.}

\item{groupIDs}{IDs for each object you want to include in the group.}
}
\value{
Returns an empty r3js group object in the form of a list.
}
\description{
This function can be used to link plot objects together into a group in order to apply
highlighting and interactive effects. See details.
}
