% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSproperties.R
\name{highlight3js}
\alias{highlight3js}
\title{Add interactive highlighting to an r3js plot object}
\usage{
highlight3js(data3js, highlight)
}
\arguments{
\item{data3js}{The data3js object}

\item{highlight}{Arguments to apply to the highlighted object}
}
\description{
r3js allows for interactive highlighting of a plot object upon mouse-over, for
example changing the color or size of a point. This can be particularly helpful
when combined with a point label which will also show on mouse-over. The
highlight parameter takes a list of arguments to apply to the highlighted
version of the object, for example if you want a point to turn blue and increase
to size 4, you could supply `highlight = list(col="blue", size=4)`.
}
