% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{legend3js}
\alias{legend3js}
\title{Add a legend to a 3js plot}
\usage{
legend3js(data3js, legend, fill, title, line_spacing = "4px")
}
\arguments{
\item{data3js}{The r3js plotting data.}

\item{legend}{A character or expression vector of length ≥ 1 to appear in the legend.}

\item{fill}{If specified, this argument will cause boxes filled with the specified colors.}

\item{title}{A character string or length-one expression giving a title to be placed at the top of the legend.}
}
\description{
Add a legend to a 3js plot
}
