% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{legend3js}
\alias{legend3js}
\title{Add a legend to an data3js object}
\usage{
legend3js(data3js, legend, fill, padding.bottom = "8px", padding.right = "8px")
}
\arguments{
\item{data3js}{The data3js object}

\item{legend}{Character vector of legend labels}

\item{fill}{If supplied the fill color of a box placed next to each label}

\item{padding.bottom}{Padding from bottom side}

\item{padding.right}{Padding from right side}
}
\description{
Add a legend to an data3js object
}
