% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSobjects.R
\name{light3js}
\alias{light3js}
\title{Add a light source to a data3js object}
\usage{
light3js(data3js, position = NULL)
}
\arguments{
\item{data3js}{The data3js object}

\item{position}{Position of the light source in x, y, z coords}
}
\description{
When no light source is provided the 3d scene is lit from the top left, this
function allows you to specify different numbers of light sources at
different positions - not yet fully implemented.
}
