% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSlines.R
\name{line3js}
\alias{line3js}
\title{Add a single straight line to a data3js object}
\usage{
line3js(data3js, x, y, z, lwd = 1, col = "black", mat = "basic", ...)
}
\arguments{
\item{data3js}{The data3js object}

\item{x}{x coords}

\item{y}{y coords}

\item{z}{z coords}

\item{lwd}{line width}

\item{col}{line color}

\item{mat}{material (see `material3js()`)}

\item{...}{further parameters to pass to `material3js()`}
}
\description{
This underlies the lines3js function.
}
