% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSbaseobjects.R
\name{line3js}
\alias{line3js}
\title{Add a line to a 3js object}
\usage{
line3js(data3js, x, y, z, lwd = 1, col = "black", mat = "basic",
  removeSelfTransparency = TRUE, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Add a line to a 3js object
}
