% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSgroupobjects.R
\name{lines3js}
\alias{lines3js}
\title{Add lines to a 3js object}
\usage{
lines3js(data3js, x, y, z, lwd = 1, col = "black", highlight,
  geometry = FALSE, ...)
}
\arguments{
\item{...}{}
}
\description{
Add lines to a 3js object
}
