% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSproperties.R
\name{material3js}
\alias{material3js}
\title{Set material properties of 3js object}
\usage{
material3js(
  mat = "phong",
  col = "black",
  opacity = NULL,
  xpd = TRUE,
  lwd = 1,
  lend = 0,
  ljoin = 0,
  dashSize = NULL,
  gapSize = NULL,
  interactive = NULL,
  moveable = NULL,
  label = NULL,
  toggle = NULL,
  dimensions = NULL,
  depthWrite = NULL,
  depthTest = NULL,
  polygonOffset = NULL,
  polygonOffsetFactor = NULL,
  polygonOffsetUnits = NULL,
  shininess = 30,
  faces = NULL,
  corners = NULL,
  rotation = NULL,
  normalise = NULL,
  poffset = NULL,
  clippingPlanes = NULL,
  frontSide = TRUE,
  backSide = TRUE,
  renderOrder = NULL,
  renderSidesSeparately = NULL,
  removeSelfTransparency = NULL,
  breakupMesh = NULL,
  ...
)
}
\arguments{
\item{mat}{Material to use for the object, one of "basic", "lambert", "phong"
or "line", see e.g.
[MeshBasicMaterial](https://threejs.org/docs/index.html#api/en/materials/MeshBasicMaterial)}

\item{col}{Color}

\item{opacity}{Opacity}

\item{xpd}{Should parts of the object outside the plot limits be shown}

\item{lwd}{Line width}

\item{lend}{Line end type (not implemented)}

\item{ljoin}{Line join type (not implemented)}

\item{dashSize}{Dash size for dashed lines}

\item{gapSize}{Gap size for dashed lines}

\item{interactive}{Is the object interactive}

\item{moveable}{Is the object moveable (not implemented)}

\item{label}{The label for the object}

\item{toggle}{Toggle button associated with the object}

\item{dimensions}{Dimensions of the object - relevant for point geometries, e.g. 3d is sphere 2d is a circle}

\item{depthWrite}{See [depthWrite](https://threejs.org/docs/index.html#api/en/materials/Material.depthWrite)}

\item{depthTest}{See [depthTest](https://threejs.org/docs/index.html#api/en/materials/Material.depthTest)}

\item{polygonOffset}{See [polygonOffset](https://threejs.org/docs/index.html#api/en/materials/Material.polygonOffset)}

\item{polygonOffsetFactor}{See [polygonOffsetFactor](https://threejs.org/docs/index.html#api/en/materials/Material.polygonOffsetFactor)}

\item{polygonOffsetUnits}{See [polygonOffsetUnits](https://threejs.org/docs/index.html#api/en/materials/Material.polygonOffsetUnits)}

\item{shininess}{Shininess of object surface}

\item{faces}{For dynamically hidden objects, the face with which it is associated, see details.}

\item{corners}{For dynamically hidden objects, the corners with which it is associated, see details.}

\item{rotation}{In place rotation of the object geometry (most relevant for points)}

\item{normalise}{Should coordinates be normalised to be with respect to axis ranges or placed according to the plotting box which has unit coordinates.}

\item{poffset}{Positional offset, the offset is relative to the plotting area size rather than axis limits}

\item{clippingPlanes}{Clipping planes to apply to the object}

\item{renderOrder}{See [renderOrder](https://threejs.org/docs/index.html#api/en/core/Object3D.renderOrder)}

\item{renderSidesSeparately}{Render the front and back side separately (not implemented)}

\item{removeSelfTransparency}{Remove self-transparency (not implemented)}

\item{breakupMesh}{Breakup the object mesh to achieve more realistic transparency rendering (not implemented)}

\item{...}{Additional parameters that are not used}

\item{doubleSide}{Should the object be rendered as double sided, see [Material Constants](https://threejs.org/docs/index.html#api/en/constants/Materials)}
}
\description{
Set material properties of 3js object
}
\details{

}
