% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSplotting.R
\name{material3js}
\alias{material3js}
\title{Set material properties of 3js object}
\usage{
material3js(mat = "phong", col = "black", opacity = NULL,
  xpd = TRUE, lwd = 1, lend = 0, ljoin = 0, dashSize = NULL,
  gapSize = NULL, interactive = NULL, moveable = NULL,
  label = NULL, toggle = NULL, dimensions = NULL,
  depthWrite = NULL, depthTest = NULL, polygonOffset = NULL,
  polygonOffsetFactor = NULL, polygonOffsetUnits = NULL,
  shininess = 30, faces = NULL, corners = NULL, rotation = NULL,
  normalise = NULL, poffset = NULL, clippingPlanes = NULL,
  doubleSide = TRUE, renderOrder = NULL,
  renderSidesSeparately = NULL, removeSelfTransparency = NULL,
  breakupMesh = NULL, ...)
}
\arguments{
\item{opacity}{}
}
\description{
Set material properties of 3js object
}
