% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSplotting.R
\name{material3js}
\alias{material3js}
\title{Set material properties of 3js object}
\usage{
material3js(mat = "phong", col = "black", opacity = 1, xpd = TRUE,
  lwd = 1, interactive, label, toggle, dimensions, depthWrite, polygonOffset,
  polygonOffsetFactor, polygonOffsetUnits, faces)
}
\arguments{
\item{opacity}{}
}
\description{
Set material properties of 3js object
}
