% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSplotting.R
\name{plot3js}
\alias{plot3js}
\title{Generic r3js plotting function}
\usage{
plot3js(
  x,
  y,
  z,
  xlim = extendrange(x),
  ylim = extendrange(y),
  zlim = extendrange(z),
  xlab = NULL,
  ylab = NULL,
  zlab = NULL,
  type = "p",
  axisline = 3,
  aspect = c(1, 1, 1),
  fixed_aspect = FALSE,
  show_plot = TRUE,
  label_axes = "xyz",
  show_grid = TRUE,
  grid_lwd = 1,
  axis_lwd = grid_lwd,
  box_lwd = grid_lwd,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Generic r3js plotting function
}
