% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSplotting.R
\name{point3js}
\alias{point3js}
\title{Add point to a 3js object}
\usage{
point3js(data3js, x, y, z, size, col, pch = 0, highlight, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Add point to a 3js object
}
