% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSplotting.R
\name{points3js}
\alias{points3js}
\title{Add points to a 3js object}
\usage{
points3js(data3js, x, y, z, size = 1, col = "black", pch = 16,
  highlight, label = NULL, toggle = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Add points to a 3js object
}
