% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSplotting.R
\name{pretty_axis}
\alias{pretty_axis}
\title{Function to generate pretty axis labels that don't go outside range of x}
\usage{
pretty_axis(x, n = 5, include_lims = TRUE)
}
\arguments{
\item{include_lims}{}
}
\description{
Function to generate pretty axis labels that don't go outside range of x
}
