% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSmethods.R
\name{print.data3js}
\alias{print.data3js}
\alias{plot.data3js}
\title{Print method for data3js objects}
\usage{
\method{print}{data3js}(x, ..., view = interactive())

\method{plot}{data3js}(x, ..., view = interactive())
}
\arguments{
\item{x}{The data3js object}

\item{...}{Additional arguments, not used}
}
\description{
The print method for the data3js object, by default plotting the htmlwidget.
}
