% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSmethods.R
\name{print.data3js}
\alias{print.data3js}
\title{Print data3js}
\usage{
\method{print}{data3js}(x, ...)
}
\arguments{
\item{x}{The data3js object}

\item{...}{Additional arguments, not used}
}
\description{
The print method for the data3js object, by default plotting the htmlwidget.
}
