% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSmethods.R
\name{print.data3js}
\alias{print.data3js}
\title{Print data3js}
\usage{
\method{print}{data3js}(data3js)
}
\arguments{
\item{data3js}{The r3js data object}
}
\description{
The print method for the data3js object, by default plotting the htmlwidget.
}
