% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r3js.R
\name{r3js}
\alias{r3js}
\title{Plot r3js data}
\usage{
r3js(data3js, width = NULL, height = NULL, elementId = NULL,
  rotation = NULL, zoom = NULL, translation = NULL,
  placeholder = NULL)
}
\description{
Creates an html widget from an r3js data object.
}
