% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r3js.R
\name{r3js}
\alias{r3js}
\title{<Add Title>}
\usage{
r3js(mapData, width = NULL, height = NULL, elementId = NULL,
  rotation = c(-45, 0, 0), zoom = 1, translation = c(0, 0, 0),
  show_rotation = FALSE)
}
\description{
<Add Description>
}
