% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSgroupobjects.R
\name{segments3js}
\alias{segments3js}
\title{Add lines segments a 3js object}
\usage{
segments3js(data3js, x, y, z, lwd = 1, col = "black", highlight, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Add lines segments a 3js object
}
