% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSobjects.R
\name{shape3js}
\alias{shape3js}
\title{Add a generic shape to an 3js plot}
\usage{
shape3js(
  data3js,
  vertices,
  faces,
  normals = NULL,
  col = "black",
  highlight,
  ...
)
}
\arguments{
\item{data3js}{The data3js object}

\item{vertices}{An nx3 matrix of 3d vertex coordinates}

\item{faces}{An nx3 matrix of indices relating to vertices that make up each triangular face}

\item{normals}{Optional nx3 matrix of normals to each vertex}

\item{col}{Shape color}

\item{highlight}{highlight attributes (see `highlight3js()`)}

\item{...}{Additional attributes to pass to `material3js()`}
}
\description{
Add a generic shape to an 3js plot
}
