% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSplotting.R
\name{sidegrid3js}
\alias{sidegrid3js}
\title{Add grid lines to an axis}
\usage{
sidegrid3js(data3js, ax, side, at, col = "grey80", dynamic = FALSE, ...)
}
\arguments{
\item{...}{}
}
\description{
Add grid lines to an axis
}
