% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSutils.R
\name{snapshot3js}
\alias{snapshot3js}
\title{Take a map snapshot}
\usage{
snapshot3js(data3js, filename, height = NULL, width = NULL, ...)
}
\arguments{
\item{filename}{The filename of the snapshot}

\item{height}{Optional height of the screenshot (in pixels), see details.}

\item{width}{Optional width of the screenshot (in pixels), see details.}

\item{...}{Further parameters to pass to \code{RacViewer()}}

\item{map}{The map data object}
}
\description{
Takes a snapshot of a map and saves it as a png
}
