% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSplotting.R
\name{sphere3js}
\alias{sphere3js}
\title{Add a sphere of defined radius to an r3js plot}
\usage{
sphere3js(data3js, x, y, z, radius, col, highlight, ...)
}
\arguments{
\item{...}{}
}
\description{
Add a sphere of defined radius to an r3js plot
}
