% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSplotting.R
\name{string3js}
\alias{string3js}
\title{Add a single text string to a 3js plot}
\usage{
string3js(data3js, x, y, z, text, size = 1, col = "black",
  type = "geometry", label = NULL, toggle = NULL, pos = c(0.5, 0.5),
  offset = c(0, 0), ...)
}
\arguments{
\item{pos}{}
}
\description{
Add a single text string to a 3js plot
}
