% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSobjects.R
\name{surface3js}
\alias{surface3js}
\title{Add a surface to an data3js object}
\usage{
surface3js(
  data3js,
  x,
  y,
  z,
  col = "black",
  mat,
  wireframe = FALSE,
  highlight,
  ...
)
}
\arguments{
\item{data3js}{The data3js object}

\item{x}{Values corresponding to rows of z, or matrix of x coordinates}

\item{y}{Values corresponding to the columns of z, or matrix of y coordinates}

\item{z}{Matrix of heights}

\item{col}{The color of the surface as either a single value, vector or matrix.}

\item{mat}{The material to use when drawing the matrix, for a solid surface the default is
"phong", for a wireframe the default is "line".}

\item{wireframe}{Logical value for if the surface should be displayed as a mesh}

\item{highlight}{highlight attributes (see `highlight3js()`)}

\item{...}{Material and texture properties. See `material3js()`}
}
\description{
This function behaves very similarly to the \code{surface3d} function in the \code{rgl}
package, although the handling of NA values is more robust in this implementation.
}
