% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JStext.R
\name{text3js}
\alias{text3js}
\title{Add text to a data3js object}
\usage{
text3js(
  data3js,
  x,
  y,
  z,
  text,
  size = 1,
  col = "inherit",
  toggle = NULL,
  type = "geometry",
  alignment = "center",
  offset = c(0, 0),
  style = list(),
  ...
)
}
\arguments{
\item{data3js}{The data3js object}

\item{x}{x coord}

\item{y}{y coord}

\item{z}{z coord}

\item{text}{text string}

\item{size}{text size}

\item{col}{text color}

\item{toggle}{associated text toggle button}

\item{type}{text type, either "geometry" or "html"}

\item{alignment}{text alignment, i.e. "left" "top" "topright"}

\item{offset}{onscreen text offset for html text x then y}

\item{style}{named list of css style attributes to apply to the html text}

\item{...}{Additional attributes to pass to `material3js()`}
}
\description{
The text added can either be as an html text object, superimposed on the scene
but moving relative to appear relative to the specified coordiantes, or an actual
geometry, which will appear in the scene, zoom and rotate with it etc.
}
