% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSplotting.R
\name{text3js}
\alias{text3js}
\title{Add text to a 3js plot}
\usage{
text3js(data3js, x, y, z, text, size = 1, col = "inherit",
  label = NULL, toggle = NULL, type = "geometry",
  alignment = "center", offset = c(0, 0), style = list(), ...)
}
\arguments{
\item{...}{}
}
\description{
Add text to a 3js plot
}
