% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3JSplotting.R
\name{triangle3js}
\alias{triangle3js}
\title{Add a triangle to an r3js plot}
\usage{
triangle3js(data3js, vertices, col = "black", highlight, label = NULL,
  toggle = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Add a triangle to an r3js plot
}
